统一日志、消息、报警与错误码描述架构升级
本次提交实现全项目日志、消息通知、报警处理的统一抽象与架构升级：
- 新增 IErrorCatalog、ErrorDescriptor、JsonErrorCatalog 及 errorcatalog.json，支持错误码到详细描述、建议的映射查询。
- AppReporter 增强，ServiceBase、ViewModel、Tools 等类统一通过 IAppReporter 处理日志、消息、报警，避免重复实现。
- SystemMessage 增加 Description、Suggestion 字段，消息内容更丰富。
- SystemContext 初始化流程调整，注入 ErrorCatalog 和 Reporter。
- 服务层与工具类重构，全部采用统一 Result/Result<T> 返回模型，错误处理与日志通知走 AppReporter。
- AxisTableInfo.xaml.cs 重构，采用 ObservableCollection 和 INotifyPropertyChanged 优化数据绑定。
- 删除领域分拆结果类型，Result<T> 增加 OkList 工厂方法，错误码枚举补充 Unknown。
整体提升代码层级清晰度、模块化、易维护性，并支持更丰富的 UI 错误展示与诊断。

+++ AM.MotionService/Base/MotionCardBase.cs
@@ -1,6 +1,4 @@
 ﻿using AM.Core.Base;
-using AM.Core.Context;
-using AM.Core.Messaging;
 using AM.Model.Common;
 using AM.Model.Interfaces.MotionCard;
 using AM.Model.MotionCard;
@@ -41,12 +39,12 @@ namespace AM.MotionService.Base
 
         protected Result Fail(MotionErrorCode code, string message)
         {
-            return base.Fail((short)code, message);
+            return base.Fail((int)code, message);
         }
 
         protected Result<T> Fail<T>(MotionErrorCode code, string message)
         {
-            return base.Fail<T>((short)code, message);
+            return base.Fail<T>((int)code, message);
         }
 
         #region 轴配置管理
